platform/libva-fake-driver: initial commit
This CL presents the initial code for a fake libva driver. Libva
(x11-libs/libva) is an API for video/image decoding/encoding
acceleration, and this package provides a Fake backend for it
for VMs and other test-related images.

Note that weirdly enough libva wants drivers to be in
  /usr/lib64/va/drivers/fake_drv_video.so  (no lib prefix)

this package installs the generated .so in /usr/lib64 as
is customary and makes a symbolic link to it.

crrev.com/c/2193152 has the ebuild file for this.

BUG=chromium:1080871
TEST=cros_workon --board=$BOARD start libva-fake-driver
with BOARD=amd64-generic and then `emerge-$BOARD -v libva-fake-driver`

Change-Id: Icc3c23f414f1a36b8e8cd85b6941819c68a08eef
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/libva-fake-driver/+/2200238
Commit-Queue: Miguel Casas <mcasas@chromium.org>
Tested-by: Miguel Casas <mcasas@chromium.org>
Reviewed-by: Jao-ke Chin-Lee <jchinlee@google.com>
Reviewed-by: Andres Calderon Jaramillo <andrescj@google.com>